Day 12: Document the fact that actual input is troll input
The community discovered that going through DFS is actually not
necessary. For the actual input, just calculate the area for the region
and the presents, and a simple comparison is sufficient. It is specially
crafted to behave that way. Hence it is a troll input because by right,
this is a NP-hard problem and DFS is needed to confirm the actual
answer, but we don't need to.

It will not work on sample input (the non-troll version), so that will
continue to use DFS.